+++ phone/src/apps/bank/components/navigation/NavigationBar.tsx
@@ -1,9 +1,9 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 
 import { Link } from 'react-router-dom';
 import { BottomNavigation, BottomNavigationAction } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
-import { Home, List, Payment } from '@mui/icons-material';
+import { Home, EnhancedEncryption } from '@mui/icons-material';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -30,8 +30,7 @@ export const NavigationBar = () => {
       className={classes.root}
     >
       <BottomNavigationAction component={Link} icon={<Home />} to="/bank" />
-      <BottomNavigationAction component={Link} icon={<Payment />} to="/bank/account" />
-      <BottomNavigationAction component={Link} icon={<List />} to="/bank/transactions" />
+      <BottomNavigationAction component={Link} icon={<EnhancedEncryption />} to="/bank/crypto" />
     </BottomNavigation>
   );
 };
