--- phone/src/apps/example/example.theme.ts
@@ -1,9 +1,0 @@
-const theme = {
-  palette: {
-    primary: {
-      main: '#f23123',
-    },
-  },
-};
-
-export default theme;
