--- phone/src/apps/example/hooks/useExampleService.ts
@@ -1,5 +1,0 @@
-export const useExampleService = () => {
-  // If you need some data that you can't get through a fetch, use services / listeners with `useNuiEvent`
-};
-
-// call this function in the Phone.tsx file.
