--- phone/src/apps/example/providers/ExampleThemeProvider.tsx
@@ -1,5 +1,0 @@
-import { createAppThemeProvider } from '@os/apps/utils/createAppThemeProvider';
-
-import exampleTheme from '../example.theme';
-
-export const ExampleThemeProvider = createAppThemeProvider(exampleTheme);
